+++ frontend/src/components/Header.vue
@@ -201,6 +201,9 @@ export default {
             if ( (this.user) && (this.user.isAdmin) ){
                 t.push({ id: 14, name: "Admin", route: `/admin`, icon: 'mdi-cog', disabled: false });
             }
+            if ( (this.user) && (this.enabledPhase >= 2) ){
+                t.push({ id: 15, name: "Search", route: `/search`, icon: 'mdi-file-search', disabled: false});
+            }
 
             return t;
         },
@@ -219,7 +222,7 @@ export default {
 
         $route (to){
             this.clearMCNotification();
-            
+
             if (to.params.id){
                 let type = (to.name === "upload_view") ? "upload" : false;
                 type = (!type && to.name === "datasets_form") ? "dataset" : type;
@@ -241,7 +244,7 @@ export default {
                     this.pendingMCMessage = m;
                 }
             }
-            
+
         },
 
         jwt(){
@@ -310,7 +313,7 @@ export default {
         },
 
         keepAlive: async function(){
-            
+
             let tok = await authServ.getToken(this.jwt);
             if (!tok || tok.error){
                 this.showError = true;
@@ -319,14 +322,14 @@ export default {
                 this.showError = false;
                 this.error = "";
             }
-            
+
         },
 
         forumApiMessage: function(event){
-            
+
             this.showNotification = false;
             //this.notificationText = ''
-            
+
             try{
                 let data = JSON.parse(event.data);
 
@@ -367,7 +370,7 @@ export default {
                     this.notificationText += "<a class='tall' href='" + url + "'>New " + this.$tc(type) + " created</a>"
                 }else if(data.comment){
                     //new comment somewhere
-                    this.notificationText += '<a class="tall" href="' + url + '">' + this.$tc("New comment on ") + this.$tc(type) 
+                    this.notificationText += '<a class="tall" href="' + url + '">' + this.$tc("New comment on ") + this.$tc(type)
                     this.notificationText += " - " + data.comment.comment + '</a>'
                 }
                 this.showNotification = true;
@@ -383,7 +386,7 @@ export default {
         },
 
         mcMessage: function(event){
-            
+
             try{
                 let data = JSON.parse(event.data);
 
@@ -438,7 +441,7 @@ export default {
     async mounted(){
         this.dark = this.useDark;
         this.$vuetify.theme.dark = this.dark
-        
+
         let urlConf = await this.$store.dispatch('config/getItem', {field: 'key', value: 'forumApiWS', def: {key: 'forumApiWS', value: ''}});
         this.forumWSUrl = urlConf.value;
         if (this.forumWSUrl !== '' && this.jwt){