+++ frontend/src/components/JsonEditor/JsonEditor.vue
@@ -1358,7 +1358,9 @@ export default{
                             }
                         }
                     }
-                    this.workingVal.resources[i].path = this.workingVal.resources[i].path.toString();
+                    if (this.workingVal.resources[i].path.includes(',')) {
+                        this.workingVal.resources[i].path = this.workingVal.resources[i].path.split(',');
+                    }
                 }
             }
         }